# backend/internal/interface/http/handlers/openapi.json
{
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "Multi-tenant project management + usage-based billing API. Bearer\nauth accepts either a user session JWT or an org API key (fbk_live_…).",
        "title": "Fluxboard API",
        "version": "1.0"
    },
    "externalDocs": {
        "description": "",
        "url": ""
    },
    "paths": {
        "/orgs/{orgId}/api-keys": {
            "get": {
                "parameters": [
                    {
                        "description": "Organization ID",
                        "in": "path",
                        "name": "orgId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List API keys",
                "tags": [
                    "api-keys"
                ]
            }
        },
        "/orgs/{orgId}/projects/{projectId}/analytics": {
            "get": {
                "parameters": [
                    {
                        "description": "Organization ID",
                        "in": "path",
                        "name": "orgId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Project ID",
                        "in": "path",
                        "name": "projectId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Start day (YYYY-MM-DD)",
                        "in": "query",
                        "name": "from",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "End day (YYYY-MM-DD)",
                        "in": "query",
                        "name": "to",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Project analytics",
                "tags": [
                    "analytics"
                ]
            }
        },
        "/orgs/{orgId}/usage": {
            "get": {
                "parameters": [
                    {
                        "description": "Organization ID",
                        "in": "path",
                        "name": "orgId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Org usage dashboard",
                "tags": [
                    "analytics"
                ]
            }
        }
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/api/v1"
        }
    ]
}